Time series (#3)
* Add workflow

* added support for alphanumeric freqencies

* first attempt at implementation - no tests added

* Added documentation for Croston

* add RMSSE

* remove commas to keep inputs in one line (autoblack)

* Adding dockerfile for codespaces to support pystan

* added dev container

* added requirements to devcontainer.

* added tests for the function get_sp_from_str

* Croston-demo (#1)

* refactored time series tests per #1892

* Delete Dockerfile

* Delete devcontainer.json

* Delete .gitignore

* replaced deleted files. (#2)

* replaced devcontainer stuff

* Removed equality checker

* Revised search grid

* Removed redundant code

* Added documentation/examples for PR

* added newline to .gitignore

* added newline to dockerfile

* change to gitignore to remove ref to devcontainer

* updated example notebook

* Delete Dockerfile

Not needed

* removed ipynb from gitignore

* Delete devcontainer.json

not needed

* Delete build_publish_ts.yml

not needed in this branch. This is only for publishing from `time_series_beta`

* Implemented change to tune grid and removed self.strictly_positive

* changed croston demo
